requirements.txt
streamlit>=1.55.0,<2.0.0
llama-cpp-python>=0.3.16,<0.4.0
faiss-cpu>=1.13.0,<2.0.0
sentence-transformers>=5.3.0,<6.0.0
pdfplumber>=0.11.9,<1.0.0
numpy>=2.4.3,<3.0.0
pillow>=12.1.1,<13.0.0
pandas>=2.3.3,<3.0.0
openpyxl>=3.1.5,<4.0.0
pyspark>=4.1.1,<5.0.0
SQLAlchemy>=2.0.48,<3.0.0
PyMySQL>=1.1.1,<2.0.0
python-dotenv>=1.2.2,<2.0.0
